requirements.txt
cyberwave
elevenlabs
mediapipe
opencv-python
python-dotenv
